turnament_table, turnament fight, avarage_skills to function

+++ gladiator/Fighter.h
@@ -10,7 +10,7 @@ public:
     Fighter();
     ~Fighter();
     std::string name;
-    int skills[3][9];
+    float skills[3][9];
     int id_count;
     void show_stats();
     void show_stats_all();
